file structure edited
