--- back/books/utils_test.py
@@ -1,4 +1,0 @@
-from utils import OpenAiAPI
-
-ai_instance = OpenAiAPI()
-print(ai_instance.get_author_info('한강'))
